requirements.txt
pyqt5
streamlit
pyyaml
python-crontab
psutil
sh
streamlit_authenticator
streamlit_float